% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot the data over map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The number of a state}

\item{year}{The year value}
}
\description{
Based on the \code{state.num} and \code{year} a summary of accidents occured in a given month and year is prepared.
The same information is filtered for a given valid state number and the result is
plotted over a state map.
}
\details{
The function plots the information over a map.
}

