% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read month and year from accident_<year>.csv}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vectorof year.}
}
\value{
This function returns a data frame of month and year as columns.
}
\description{
The function takes year vector as input and reads data for the specific year csv files and
returns a data frame of MONTH and year found in all of the year range csv files.
It reports with an error if any invalid year is specified.
}
\details{
The function contructs a new filename.
}

