% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summary of accidents}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector}
}
\value{
This function returns a data frame of month, year and total number of incedents as columns.
}
\description{
This function summarizes the accident data like the number of cases permonth in a given year.
}
\details{
The function contructs a new filename.
}
\examples{
\dontrun{
fars_summarize_years(2013:2016)
}

}

