% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Construct a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numerical input representing a year.}
}
\value{
This function returns a Character sting of new file name.
}
\description{
The function takes year as input and contructs a new file name by attaching the year \code{year} value
The expected result should be like accident_<year>.csv.bz2
}
\details{
The function contructs a new filename.
}
\examples{
\dontrun{
makefilename(2013)
}
}

